+++ package.json
@@ -22,6 +22,7 @@
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "react-router-dom": "^6.22.3",
     "typescript": "^5.2.2",
     "vite": "^5.1.6"
   }
