+++ src/pages/pokemonDetail/PokemonDetail.tsx
@@ -1,20 +1,19 @@
-'use client';
 
 import { useState, useEffect } from 'react';
 import styles from "./pokemonDetail.module.css";
 import Pokemon from "../../types/Pokemon";
 import {APIPokemonService} from "../../services/API";
+import { useParams, useNavigate  } from 'react-router-dom'
 
 export default function Page () :JSX.Element{
     const [data, setData] = useState<Pokemon>();
-
-    const pokemonName = "ditte";
-
+    const {pokemonName} = useParams();
+    const navigate = useNavigate();
     const ApiService = new APIPokemonService();
+
     useEffect(() => {
         ApiService.APIcallPokemon(setData, "/" + pokemonName);
         }, [])
-
 
     function PokemonDetail () {
         if (typeof data === "object"){
@@ -27,7 +26,7 @@ export default function Page () :JSX.Element{
                         <h2 className={styles.pokemonName}>{data.name}</h2>
                         <img src={pokemonImgUrl} alt={data.name}></img>
                     </div>
-                    <button className={styles.homeButton}>Home</button>
+                    <button className={styles.homeButton} onClick={()=>navigate("/")}>Home</button>
                 </div>
             )
         }
